ver1.0

+++ Karalog/Model/Firebase/ListFirebase.swift
@@ -42,67 +42,88 @@ class ListFirebase: ObservableObject {
     
     //MARK: - Get
     
-    //listsを取得
-    func getList(completionHandler: @escaping (Any) -> Void) {
-        userRef.getDocument() { (document, err) in
-            if let _document = document, _document.exists{
+    
+    // リストを取得する関数
+    func getList(completionHandler: @escaping (Bool) -> Void) {
+        getOrder() {_ in
+            self.getRandomList() { collection in
+                guard let collection else { return }
+                for document in collection.documents {
+                    do {
+                        let listName = try document.data(as: ListName.self)
+                        let id = document.documentID
+                        
+                        self.listImageRef = self.storageRef.child("images/user/\(String(describing: self.userID))/list/\(id).jpeg")
+                            
+                        self.listImageRef.getData(maxSize: 1 * 1024 * 1024) { data, error in
+                            if let error = error {
+                                print("Error getting list image: \(error)")
+                                completionHandler(false)
+                                return
+                            }
+                                
+                            
+                            if let data = data {
+                                manager.lists.append(Lists(listName: listName.listName, listImage: data, id: id))
+                            }
+                                
+                            if manager.lists.count == collection.documents.count + 2 {
+                                var preList: [Lists] = []
+                                for i in manager.user.listOrder {
+                                    guard let list = manager.lists.first(where: {$0.id!.contains(i)}) else { continue }
+                                    preList.append(list)
+                                }
+                                manager.lists = material.initialListData()
+                                manager.lists += preList
+                                
+                                // リストデータの取得が完了したことを completionHandler で通知
+                                completionHandler(true)
+                            }
+                        }
+                    } catch {
+                        print(error)
+                        completionHandler(false)
+                    }
+                }
+                manager.lists = material.initialListData()
+                completionHandler(true)
+            }
+        }
+    }
+
+    
+    
+    func getOrder(completionHandler: @escaping (Bool) -> Void) {
+        userRef.getDocument() { (document, error) in
+            if let document = document, document.exists {
                 do {
-                    let user = try _document.data(as: User.self)
+                    let user = try document.data(as: User.self)
                     manager.user.listOrder = user.listOrder
                 } catch {
+                    print("Error getting user data: \(error)")
+                    completionHandler(false)
                     return
                 }
                 
-                print("getting listOrder")
-                
-                self.listRef.getDocuments() { (collection, err) in
-                    if let _err = err {
-                        print("error getting list: \(_err)")
-                        
-                    }else{
-                        print("getting list")
-                        manager.lists = []
-                        
-                        for document in collection!.documents {
-                            do{
-                                let listName = try document.data(as: ListName.self)
-                                var listImage: Data!
-                                let id = document.documentID
-                                
-                                self.listImageRef = self.storageRef.child("images/user/\(String(describing: self.userID))/list/\(id).jpeg")
-                                let storagePath = "gs://karalog-39e53.appspot.com/images/user/\(String(describing: self.userID))/list/\(id).jpeg"
-                                
-                                self.listImageRef.getData(maxSize: 1 * 1024 * 1024) { data, err in
-                                    if let err {
-                                        print("Error get list image: \(err)")
-                                        return
-                                    } else {
-                                        listImage = data
-                                        manager.lists.append(Lists(listName: listName.listName, listImage: listImage, id: id))
-                                    }
-                                }
-                                
-                            }catch{
-                                print(error)
-                            }
-                        }
-                        var preList: [Lists] = []
-                        for i in manager.user.listOrder {
-                            guard let list = manager.lists.first(where: {$0.id!.contains(i)}) else { continue }
-                            preList.append(list)
-                        }
-                        manager.lists = material.initialListData()
-                        manager.lists += preList
-                        print(manager.lists)
-                        completionHandler(true)
-                    }
-                }
-                
+                print("Getting listOrder")
+                completionHandler(true)
             } else {
                 print("Error getting listOrder")
+                completionHandler(false)
             }
-            
-            
+        }
+    }
+    
+    func getRandomList(completionHandler: @escaping (QuerySnapshot?) -> Void) {
+        self.listRef.getDocuments() { (collection, error) in
+            if let error = error {
+                print("Error getting list: \(error)")
+                completionHandler(nil)
+                return
+            }
+            completionHandler(collection)
+            print("Getting list")
+            manager.lists = material.initialListData()
         }
     }
     
@@ -115,7 +136,8 @@ class ListFirebase: ObservableObject {
                     print("Error getting wanna list: \(String(describing: _err))")
                     continuation.resume(returning: nil)
                 } else {
-                    for document in collection!.documents {
+                    guard let collection else { return }
+                    for document in collection.documents {
                         
                         let name = document.data()[UserRef.WannaListRef.musicName.rawValue] as! String
                         let artist = document.data()[UserRef.WannaListRef.artistName.rawValue] as! String
@@ -175,7 +197,7 @@ class ListFirebase: ObservableObject {
     }
     
     //wannaListを追加
-    func addWanna(musicName: String, artistName: String, musicImage: String) {
+    func addWanna(musicName: String, artistName: String, musicImage: String, completionHandler: @escaping (Any) -> Void) {
         wannaRef.addDocument(data: [
             UserRef.WannaListRef.musicName.rawValue: musicName,
             UserRef.WannaListRef.artistName.rawValue: artistName,
@@ -196,32 +218,33 @@ class ListFirebase: ObservableObject {
     
     //listを削除
     func deleteList(indexPath: IndexPath, completionHandler: @escaping (Any) -> Void) {
-        let listID = manager.lists[indexPath.row].id!
-        listImageRef = storageRef.child("images/user/\(String(describing: userID))/list/\(listID).jpeg")
-        let storagePath = "gs://karalog-39e53.appspot.com/images/user/\(String(describing: userID))/list/\(listID).jpeg"
-        listImageRef = storage.reference(forURL: storagePath)
-        listRef.document(listID).delete() { err in
-            if let _err = err {
-                print("error removing music: \(_err)")
-            }else{
-                print("music successfully removed")
-                self.userRef.updateData([
-                    UserRef.listOrder.rawValue: FieldValue.arrayRemove([listID])
-                ]){err in
-                    if let _err = err {
-                        print("Error deleting music order: \(_err)")
-                    }else{
-                        print("music order successfully deleted")
-                        manager.lists.remove(at: indexPath.row)
-                        
-                    }
-                    
-                    self.listImageRef.delete { error in
-                        if let error {
-                            print("Error delete list image:", error)
+        if let listID = manager.lists[indexPath.row].id {
+            listImageRef = storageRef.child("images/user/\(String(describing: userID))/list/\(listID).jpeg")
+            let storagePath = "gs://karalog-39e53.appspot.com/images/user/\(String(describing: userID))/list/\(listID).jpeg"
+            listImageRef = storage.reference(forURL: storagePath)
+            listRef.document(listID).delete() { err in
+                if let _err = err {
+                    print("error removing music: \(_err)")
+                }else{
+                    print("music successfully removed")
+                    self.userRef.updateData([
+                        UserRef.listOrder.rawValue: FieldValue.arrayRemove([listID])
+                    ]){err in
+                        if let _err = err {
+                            print("Error deleting music order: \(_err)")
+                        }else{
+                            print("music order successfully deleted")
+                            manager.lists.remove(at: indexPath.row)
+                            
                         }
+                        
+                        self.listImageRef.delete { error in
+                            if let error {
+                                print("Error delete list image:", error)
+                            }
+                        }
+                        completionHandler(true)
                     }
-                    completionHandler(true)
                 }
             }
         }
